Step 2 of Two Pointers

+++ CodeForces/CF_NumberOfEqual.cpp
@@ -4,11 +4,11 @@ using namespace std;
 #define eb emplace_back
  
 void solve() {
-    int n, m; cin >> n >> m;
-    vector<int> a(n), b(m);
-    for (int i = 0; i < n; i++) cin >> a[i];
-    for (int i = 0; i < m; i++) cin >> b[i];
-    int i = 0, j = 0, result = 0, curr_value = INT_MIN, count_a = 0, count_b = 0;
+    ll n, m; cin >> n >> m;
+    vector<ll> a(n), b(m);
+    for (ll i = 0; i < n; i++) cin >> a[i];
+    for (ll i = 0; i < m; i++) cin >> b[i];
+    ll i = 0, j = 0, result = 0, curr_value = INT_MIN, count_a = 0, count_b = 0;
     while(i < n && j < m) { 
         if (a[i] == b[j]) {
             curr_value = a[i];
